Try to fix code snippet highlight
https://stackoverflow.com/questions/55308142/why-do-i-get-a-double-frame-around-markdown-code-block-on-jekyll-site

For anyone experiencing this issue still (like me) and looking for the actual bugfix since the OP never actually said how they fixed it here you go:

In the _sass/_highlights.scss file you simply need to replace .highlight with pre.highlight. It appears that some styling can be applied twice if this is not specified. I also had a entry in pre.highlight{...} that I changed from overflow: scroll; to overflow: auto; in order to hide the scroll bars if they are not necessary.

BEFORE:

.highlight{
      ...
          overflow: scroll;
              ...
}
AFTER:

pre.highlight{
      ...
          overflow: auto;
              ...
}
It appears the original issue was with some Jekyll templates that people are running into still. I found the answer from this SO answer which referenced this thread if anyone wants more info.